Refactor (#38)
* Add refactoring of pbc tests

Make them calcualte the tests for 7 frames of the same trajectory instead
of 1.

* Refactor Gyration test

Make it into a proper test in the first place. Adds comparison to
gmx gyrate -p yes (mass weighted)
Add gyration test against moments of inertia

* Garbage commit, added some keywords to histogram plotting long forgotten

* Forgot to actually commit test_gyr two commits ago..

* Refactor the cluster condensation tests

The test for cluster condenstation not specific to pbc finish
within tolerance of one molecule detected or not detected.

* Remove *.txt and *.out files from tests/cluster

* Remove unit_test_....py

Was very badly done -- easier to just do new input unit test
(low priority really).

* Refactor main.py test. Last refactored test.

This had an analysis of an atomistic trajectory.

* Delete superfluous trajectories. Add Readme.

Readme for clustercode/tests/cluster/files.

* Add clustering tests.

test_clustering.py for checking different cluster_analysis() keywords
so far static vs. dynamic size comparison of a micellising system.

* Add more clustering tests

* Add further gyration tests.

Add xfail test which checks if we get the same gyration
radius if we look for only the gyration radius of the
alkane beads in a trajectory.

* Add unwrapping test

Test if the COG after unwrapping is the same than unwrapping with
gmx trjconv -f ... -pbc cluster

* Refactor ClusterEnsemble.py. Copy code into regions

* Copy past code into ClusterSearch.py - no functional relationship yet

* Refactor cluster_analysis stuff into ClusterSearch class

* Refactor unwrap_cluster into it's own class

* Refactor condensed_ions()

* Refactor gyration methods into Gyration class

* Change UnwrapCluster and CondensedIons init fct

so they don't need the universe to be initialised

* Add angle/distance distribution tests

* Bug upload

a lot of tests were passing because the generator contraption
wasn't working. So whenever I ran through a cluster list twice,
it wasn't running through the list anymore and hence no
assert statements got triggered

* Bugfix generator

Make cluster_list a property of ClusterEnsemble returning
cluster_list of ClusterSearch instance (which is a generator).
It's not pretty but it works and tests pass.

* Refactor xyz_distribution() to Distribution:

* Refactor _gyration_tensor

And add tests for it

* Refactor sort_eig() and add tests in Gyration

* Add tests to consolidate gyration() and rgyr()

* Add more detail to gyr indepth tests

make the perfect cylinder also perfect wrt. mass and add comments
to all tests

* Refactor gyration()

add pca and mass keyword and functionality

* Refactor intertia_tensor() and test rgyr()